refactor(service.base): move AbstractTransactionalService to base package and update implementations
Relocated AbstractTransactionalService from com.gabrielosorio.gestor_inteligente.service to com.gabrielosorio.gestor_inteligente.service.base package. Updated import statements in 6 service implementations (CheckoutMovementServiceImpl, ProductServiceImpl, SaleCheckoutMovementServiceImpl, etc.) to reference the new package location. Impact: improves code organization by centralizing base classes in a dedicated package, eliminates cross-package dependencies, and establishes a clear separation between base abstractions and concrete implementations.
