+++ src/main/java/com/gabrielosorio/gestor_inteligente/service/base/AbstractTransactionalService.java
@@ -1,4 +1,4 @@
-package com.gabrielosorio.gestor_inteligente.service;
+package com.gabrielosorio.gestor_inteligente.service.base;
 
 import com.gabrielosorio.gestor_inteligente.repository.base.Repository;
 import com.gabrielosorio.gestor_inteligente.repository.strategy.base.TransactionalStrategy;
